+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to this project will be documented in this file.
 - Add "Filter by keywords in the message body" feature (#653)
 - Upgrade to HiDPI screenshots in the documentation
 - Add support for the $JSON placeholder in webhook payloads
+- Add ping endpoints for "log" events
 
 ### Bug Fixes
 - Fix the display of ignored pings with non-zero exit status